+++ NEWS.md
@@ -10,6 +10,9 @@
     list by the user, or the `generateSegments` function whose arguement is an **sf** object. 
 -   The `formAssay` function now accepts only segments without the user manually generating coordinates 
     (or centroids).
+-   The `name` argument in functions like `vrImages`, `vrCoordinates` and `vrSegments` is replaced
+    with `spatial` which is the name of the spatial coordinate system of the assay, spatial object. 
+-   The `formImage` function is replaced with `formSpatial` to avoid confusion. 
     
 # VoltRon 0.2.6